+++ addons/sys_logi/cfg/vehicle/MSF_LogisticsTerminal.hpp
@@ -1,0 +1,10 @@
+class ItemGPS;
+class MSF_LogisticsTerminal : ItemGPS
+{
+    displayName = "Logistics Terminal";
+    author = "Zeik_Tuvai";
+    descriptionShort = "Enables accessing the Logistics Terminal.";
+    hiddenSelectionsTextures[] = {"\A3\Drones_F\Weapons_F_Gamma\Items\data\UAV_controller_rgr_co.paa"};
+    model = "\a3\Drones_F\Weapons_F_Gamma\Items\UAV_controller_F";
+    picture = "\A3\Drones_F\Weapons_F_Gamma\Items\data\UI\gear_UAV_controller_rgr_CA.paa";
+};