+++ addons/sys_logi/config.cpp
@@ -23,21 +23,50 @@ class CfgVehicles
 	#include "cfg\vehicle\MSF_Supply_Depot.hpp"
 	#include "cfg\vehicle\MSF_VicTerminal.hpp"
 
-	// class Man;
-	// class CAManBase : Man
-	// {
-	// 	class ACE_SelfActions
-	// 	{
-	// 		class MSF_LogiAction
-	// 		{
-	// 			displayName = "Logistics";
-	// 			condition = "count (localNamespace getVariable 'MSF_Local' getOrDefault ['Logi_Items', []]) > 0";
-	// 			exceptions[] = {};
-	// 			icon = "a3\ui_f\data\gui\cfg\communicationmenu\supplydrop_ca.paa";
-	// 			insertChildren = "params [""_target"", ""_player"", ""_params""]; [_target, _player, _params] call MSF_Logi_fnc_GetLogiItemsMenu;";
-	// 		};
-	// 	};
-	// };
+	class Man;
+	class CAManBase : Man
+	{
+		class ACE_SelfActions
+		{
+			class MSF_Logistics
+			{
+				displayName = "Logistics Terminal";
+				icon = "\A3\Drones_F\Weapons_F_Gamma\Items\data\UI\gear_UAV_controller_rgr_CA.paa";
+				condition = "[player, 'MSF_LogisticsTerminal', false] call BIS_fnc_hasItem";
+				exceptions[] = {};
+
+				class MSF_LogisticsMenu
+				{
+					displayName = "Logistics Management";
+					exceptions[] = {};
+					icon = MSF_ICON;
+					condition = "[player, 'Engineer'] call MSF_fnc_CheckPlayerTrait";
+					statement = "[] call MSF_UI_fnc_DisplayLogiMain";
+				};
+				class MSF_LogisticsSupport
+				{
+					displayName = "Logistics Support Channel";
+					condition = "true";
+					exceptions[] = {};
+					icon = "a3\ui_f\data\gui\cfg\communicationmenu\supplydrop_ca.paa";
+					statement = "[] call MSF_UI_fnc_DisplayLogiSupport";
+				};
+				class MSF_VicLoadouts
+				{
+					displayName = "Vehicle Loadouts";
+					exceptions[] = {};
+					icon = "a3\soft_f_epc\truck_03\data\ui\truck_03_ammo_ca.paa";
+					condition = "[player, 'Engineer'] call MSF_fnc_CheckPlayerTrait";
+					statement = "[] call MSF_UI_fnc_DisplayVicLoadouts;";
+				};
+			};
+		};
+	};
+};
+
+class CfgWeapons
+{
+	#include "cfg\vehicle\MSF_LogisticsTerminal.hpp"
 };
 
 class CfgHints